Add Rush Hour Puzzle solver (task B)
I've solved task B, which is using the A* algorithm to solve the Rush
Hour Puzzle. At the moment the GUI is pretty bad, and will print a new
window for every move the AI makes (so you have to exit the window each
time). This needs to be fixed. The AI can solve the easy and medium
board perfectly. It also solves the hard board, but there is something
weird going on with the GUI, showing two green squares constantly.
The AI can not solve the expert board yet.
I will look into this and fix a new GUI that writes the changes
in the same window.