Add working GUI
The GUI is now working properly. The problem was that the draw_path
method took in a string with linebreaks. Therefore, '12' was evaluated
to '1' and '2', messing up the colors. It now takes in the state, and
draws the board of that state accordingly.

One still has to close each window though. I'll see if I have the
time/bother to fix this, but it actually works quite well. It is sort of
like pressing a button to go to the next move. The problem is that the
screen refreshes, sometimes making it hard to see what move was made.